[SEARCH] 검색화면, 결과화면 분리
