+++ app/src/main/res/layout/fragment_search.xml
@@ -89,126 +89,4 @@
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 
-    <!--검색 요청 후 결과 화면-->
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/search_result"
-        android:visibility="gone"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/result_logo_iv"
-            android:layout_width="80dp"
-            android:layout_height="30dp"
-            android:src="@drawable/splash_icon3"
-            android:layout_marginTop="30dp"
-            android:layout_marginStart="20dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"/>
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/search_bar_1"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_marginStart="10dp"
-            android:layout_marginEnd="20dp"
-            app:layout_constraintStart_toEndOf="@id/result_logo_iv"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="@id/result_logo_iv"
-            app:layout_constraintTop_toTopOf="@id/result_logo_iv" >
-            <ImageView
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:src="@drawable/search_bg"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent" />
-            <EditText
-                android:id="@+id/result_et"
-                android:layout_width="200dp"
-                android:layout_height="match_parent"
-                android:hint="원하는 놀거리를 입력해주세요"
-                android:background="@null"
-                android:textSize="10dp"
-                android:textColorHint="@color/hint"
-                android:layout_marginStart="20dp"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent" />
-            <ImageView
-                android:id="@+id/search_image_bt_1"
-                android:layout_width="17dp"
-                android:layout_height="17dp"
-                android:layout_margin="10dp"
-                android:src="@drawable/search_icon"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <ImageView
-            android:id="@+id/search_home_ic_iv"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:src="@drawable/search_home_icon"
-            android:visibility="gone"
-            android:layout_marginStart="10dp"
-            app:layout_constraintStart_toEndOf="@id/result_logo_iv"
-            app:layout_constraintBottom_toBottomOf="@id/result_logo_iv"
-            app:layout_constraintTop_toTopOf="@id/result_logo_iv" />
-
-        <ImageView
-            android:id="@+id/search_back_iv"
-            android:layout_width="25dp"
-            android:layout_height="25dp"
-            android:src="@drawable/back_arrow"
-            app:tint="@color/black"
-            android:layout_marginStart="10dp"
-            android:layout_marginTop="20dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/result_logo_iv" />
-
-        <ImageView
-            android:id="@+id/result_menu_iv"
-            android:layout_width="25dp"
-            android:layout_height="25dp"
-            android:src="@drawable/search_menu_list"
-            android:layout_marginEnd="30dp"
-            app:layout_constraintTop_toTopOf="@id/search_back_iv"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <!--리사이클러뷰가 들어갈 공간-->
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_marginTop="20dp"
-            android:layout_marginBottom="10dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/search_back_iv">
-            <!--나열형-->
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/result_list_rv"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:visibility="visible"
-                tools:listitem="@layout/item_recommend_1"
-                android:orientation="vertical"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent" />
-            <!--블록형-->
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/result_block_rv"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:visibility="gone"
-                tools:listitem="@layout/item_recommend_2"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
